+++ docs/javadocs/com/meti/asset/ByteAsset.html
@@ -101,7 +101,7 @@
                        title="class in com.meti.asset">com.meti.asset.Asset</a></li>
                 <li>
                     <ul class="inheritance">
-                        <li>com.meti.asset.ByteAsset</li>
+                        <li>com.meti.module.bytes.ByteAsset</li>
                     </ul>
                 </li>
             </ul>
@@ -120,7 +120,7 @@
 extends <a href="../../../com/meti/asset/Asset.html" title="class in com.meti.asset">Asset</a></pre>
                 <dl>
                     <dt><span class="seeLabel">See Also:</span></dt>
-                    <dd><a href="../../../serialized-form.html#com.meti.asset.ByteAsset">Serialized Form</a></dd>
+                    <dd><a href="../../../serialized-form.html#com.meti.module.bytes.ByteAsset">Serialized Form</a></dd>
                 </dl>
             </li>
         </ul>
