+++ docs/javadocs/com/meti/asset/DefaultBuilder.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.asset.DefaultBuilder</li>
+                <li>com.meti.module.bytes.ByteBuilder</li>
             </ul>
         </li>
     </ul>
