+++ docs/javadocs/com/meti/client/editor/TextEditor.html
@@ -102,7 +102,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.client.editor.TextEditor</li>
+                <li>com.meti.module.text.TextEditor</li>
             </ul>
         </li>
     </ul>
