+++ docs/javadocs/serialized-form.html
@@ -96,10 +96,10 @@
                         </li>
                     </ul>
                 </li>
-                <li class="blockList"><a name="com.meti.asset.ByteAsset">
+                <li class="blockList"><a name="com.meti.module.bytes.ByteAsset">
                     <!--   -->
                 </a>
-                    <h3>Class <a href="com/meti/asset/ByteAsset.html" title="class in com.meti.asset">com.meti.asset.ByteAsset</a>
+                    <h3>Class <a href="com/meti/asset/ByteAsset.html" title="class in com.meti.asset">com.meti.module.bytes.ByteAsset</a>
                         extends <a href="com/meti/asset/Asset.html" title="class in com.meti.asset">Asset</a> implements
                         Serializable</h3>
                 </li>
