+++ src/main/java/com/meti/module/bytes/ByteBuilder.java
@@ -1,0 +1,31 @@
+package com.meti.module.bytes;
+
+import com.meti.asset.Asset;
+import com.meti.asset.AssetBuilder;
+
+import java.io.File;
+import java.io.FileReader;
+import java.io.IOException;
+import java.util.ArrayList;
+
+public class ByteBuilder implements AssetBuilder {
+
+    @Override
+    public Asset build(File file) throws IOException {
+        FileReader reader = new FileReader(file);
+
+        //might be a bit memory intensive
+        ArrayList<Integer> integerList = new ArrayList<>();
+        int b;
+        while ((b = reader.read()) != -1) {
+            integerList.add(b);
+        }
+
+        int[] array = new int[integerList.size()];
+        for (int i = 0; i < array.length; i++) {
+            array[i] = integerList.get(i);
+        }
+
+        return new ByteAsset(array);
+    }
+}
