+++ src/main/java/com/meti/module/text/TextAsset.java
@@ -1,0 +1,6 @@
+package com.meti.module.text;
+
+import com.meti.asset.Asset;
+
+public class TextAsset extends Asset {
+}
