+++ src/main/java/com/meti/module/text/TextChange.java
@@ -1,0 +1,4 @@
+package com.meti.module.text;
+
+public class TextChange {
+}
