+++ src/main/resources/fxml/client/editor/TextEditor.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.TextArea?>
 <?import javafx.scene.layout.*?>
 <AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.111"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.meti.client.editor.TextEditor">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.meti.module.text.TextEditor">
   <children>
     <ScrollPane AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0"
       AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
